style: rename directory
